requirements.txt
clarifai_grpc==9.7.3
langchain==0.0.274
Pillow==9.2.0
python-dotenv==1.0.0
streamlit==1.25.0
clarifai
